# Major Performance Boost in `line_col` Function with `bytecount` Integration